Fixed linux compilation

+++ src/EngineWrap.cpp
@@ -99,7 +99,7 @@ Napi::Value EngineWrap::GetCameras(const Napi::CallbackInfo& info)
     Napi::Env env = info.Env();
     std::vector<Camera*> cameras = this->engine->getCameras();
     Napi::Array arr = Napi::Array::New(env, cameras.size());
-    for (int i = 0; i < cameras.size(); i++)
+    for (size_t i = 0; i < cameras.size(); i++)
         arr.Set(i, CameraWrap::NewInstance(env, cameras[i]));
     return arr;
 }
@@ -118,7 +118,7 @@ Napi::Value EngineWrap::GetPlugins(const Napi::CallbackInfo& info)
     PluginProvider& pluginProvider = this->engine->getPluginProvider();
     std::vector<PluginDescriptor> plugins = pluginProvider.getPlugins();
     Napi::Array arr = Napi::Array::New(env, plugins.size());
-    for (int i = 0; i < plugins.size(); i++)
+    for (size_t i = 0; i < plugins.size(); i++)
         arr.Set(i, PluginWrap::NewInstance(env, plugins[i]));
     return arr;
 }
